syncdeletes: start re-architecture to use mythtv urls instead of python libraries
The python libraries did not run (crashed) so this work effort is to
convert the program to use the mythtv urls instead of the python libraries.
Taking alot of learnings from the mytht2kodi scripts and appyign them here.
Have copied and pasted the Recording class for now.  Hope to create a
mythtv library that all of these scripts can use later.