myth2kodi: Add -f option to use filename instead of programid in some cases
For non season/episode based shows, sometimes the guide (and hence mythdb)
contains the same programid for all showings. This occurs most often in
daily talk shows which will never REALLY be reruns. An example is the
nightly news. For these shows, we cannot rely on programid as being
unique, so use the filename instead.